requirements.txt
Django>=5.0,<6.0
django-easy-tenants
